ref: fold corrected_pvalue into io.py and drop the single-function stats module
